Extended and cleaned up support for abbreviation lists.
Previously abbreviation lists processing worked on equality of whitespace
separated substrings in a node, e.g. "one example, i.e. ", where an
abbreviation was expected to be surrounded by blanks or at the start/end
of a text node.

Now abbreviation matching works on substring level, i.e. when an
abbreviation is defined as "&c" it will be found in a node like "This is a
example &c." - the non-matching parts of the substring "&c.", i.e. the
dot, will be processed further (and tagged as <pc>.</pc>)

- Added support for inline-abbreviation lists in the parameter file
- added support for an URL to fetch abbreviation list from (was hard coded
  before)
- updated make_xsl accordingly
- added test case for abbrevations