bugfix to implentation of floating-blocks feature:
text nodes inside of ignored elements are not to be tokenized. yet, this was checked via the icltt:textvalue() function (in lib-bdr.xsl, line 333), which, however, returns empty also for floating-blocks. Thus in cases, in which a non-ignored element contained _only_ a floating block, the floating block did not get tokenized.

fix: adding function icltt:is-ignored() to differentiate between ignored elements and floating blocks (plus helpers for generating the relevant templates in the wrapper XSL)