fixing #2
By adding an XPath parser to the profile processor stylesheet we can
make sure that a) any <expression> element actually contains a valid
XPath 2.0 expression and b) there is a namespace element defining the
URI for any QName prefix. <expression> elements which do not conform to
those rules are omitted with a non-terminating message.